scripts/api_v2_migration/requirements.txt
python-dotenv==1.1.0
pymilvus==2.5.5
pandas==2.2.3
